AI/requirements.txt
streamlit
requests
pandas
numpy
scikit-learn
joblib
matplotlib
pillow
pyttsx3
